feat: use cv section for clear past roles communication
