--- app/sections/Skills/Skills.tsx
@@ -1,11 +1,0 @@
-import { SkillsCv } from './SkillsCv'
-import { SkillsSocial } from './SkillsSocial'
-import { SkillsTech } from './SkillsTech'
-
-export const Skills = () => (
-  <>
-    <SkillsTech />
-    <SkillsSocial />
-    <SkillsCv />
-  </>
-)
