--- app/sections/Skills/SkillsCv.tsx
@@ -1,316 +1,0 @@
-import { DocumentIcon } from '@heroicons/react/24/outline'
-import { ChevronDownIcon } from '@heroicons/react/24/solid'
-import Image from 'next/image'
-import { useEffect, useRef, useState } from 'react'
-
-import { GridItem } from '@/app/components/Grid'
-
-import { SkillsWrapper } from './SkillsWrapper'
-
-type Technology = {
-  name: string
-  icon: string
-  level: 'expert' | 'advanced' | 'intermediate'
-}
-
-type Achievement = {
-  title: string
-  description: string
-  impact: string
-}
-
-type CareerPhaseProps = {
-  title: string
-  period: string
-  type: 'freelance' | 'employed'
-  description: string
-  technologies: Technology[]
-  achievements: Achievement[]
-  customerBenefits: string[]
-}
-
-const TechnologyBadge = ({ name, icon, level }: Technology) => (
-  <div className="flex items-center gap-2 rounded-full border border-gray-200 bg-white px-3 py-1.5">
-    <div className="relative h-5 w-5">
-      <Image fill src={icon} alt={`${name} icon`} className="object-contain" />
-    </div>
-    <span className="text-sm font-medium">{name}</span>
-    <span
-      className={`rounded-full px-2 py-0.5 text-xs ${
-        level === 'expert'
-          ? 'bg-blue-100 text-blue-700'
-          : level === 'advanced'
-            ? 'bg-green-100 text-green-700'
-            : 'bg-gray-100 text-gray-700'
-      }`}
-    >
-      {level}
-    </span>
-  </div>
-)
-
-const CareerPhase = ({
-  title,
-  period,
-  type,
-  description,
-  technologies,
-  achievements,
-  customerBenefits,
-}: CareerPhaseProps) => {
-  const [isExpanded, setIsExpanded] = useState(false)
-  const contentRef = useRef<HTMLDivElement>(null)
-  const [contentHeight, setContentHeight] = useState(0)
-
-  useEffect(() => {
-    if (contentRef.current) {
-      setContentHeight(isExpanded ? contentRef.current.scrollHeight : 0)
-    }
-  }, [isExpanded])
-
-  return (
-    <div
-      className={`relative rounded-lg border-2 p-6 hover:cursor-pointer ${
-        type === 'freelance'
-          ? 'border-blue-500 bg-black'
-          : 'border-gray-500 bg-black'
-      }`}
-      onClick={() => setIsExpanded(!isExpanded)}
-    >
-      <div className="absolute top-1/2 -left-3 h-6 w-6 -translate-y-1/2 rounded-full border-2 border-white bg-white" />
-
-      <div className="mb-4 flex items-start justify-between">
-        <div>
-          <h3 className="mb-2 text-xl font-bold text-white">{title}</h3>
-          <p className="mb-2 text-sm text-white">{period}</p>
-        </div>
-        <button
-          onClick={(e) => {
-            e.stopPropagation()
-            setIsExpanded(!isExpanded)
-          }}
-          className="rounded-full border-1 border-transparent p-1 transition-colors hover:cursor-pointer hover:border-white"
-        >
-          <div
-            className={`transform transition-transform duration-300 ${isExpanded ? 'rotate-180' : ''}`}
-          >
-            <ChevronDownIcon className="h-5 w-5 text-white" />
-          </div>
-        </button>
-      </div>
-
-      <p className="mb-4 text-white">{description}</p>
-
-      <div
-        className="overflow-hidden transition-all duration-300 ease-in-out"
-        style={{ height: `${contentHeight}px` }}
-      >
-        <div
-          ref={contentRef}
-          className="mt-4 space-y-6 border-t border-gray-200 pt-4"
-        >
-          {/* Technologies Section */}
-          <div>
-            <h4 className="mb-3 text-sm font-semibold text-gray-400">
-              Technologien & Expertise
-            </h4>
-            <div className="flex flex-wrap gap-2">
-              {technologies.map((tech) => (
-                <TechnologyBadge key={tech.name} {...tech} />
-              ))}
-            </div>
-          </div>
-
-          {/* Achievements Section */}
-          <div>
-            <h4 className="mb-3 text-sm font-semibold text-gray-400">
-              Wichtige Erfolge
-            </h4>
-            <div className="space-y-3">
-              {achievements.map((achievement, index) => (
-                <div
-                  key={index}
-                  className="rounded-lg border border-gray-700 bg-gray-900 p-3"
-                >
-                  <h5 className="font-medium text-blue-400">
-                    {achievement.title}
-                  </h5>
-                  <p className="mt-1 text-sm text-gray-300">
-                    {achievement.description}
-                  </p>
-                  <p className="mt-2 text-sm text-green-400">
-                    <span className="font-medium">Impact:</span>{' '}
-                    {achievement.impact}
-                  </p>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* Customer Benefits Section */}
-          <div>
-            <h4 className="mb-3 text-sm font-semibold text-gray-400">
-              Ihr Vorteil
-            </h4>
-            <ul className="space-y-2">
-              {customerBenefits.map((benefit, index) => (
-                <li
-                  key={index}
-                  className="flex items-start gap-2 text-sm text-gray-300"
-                >
-                  <span className="mt-1 text-blue-400">•</span>
-                  <span>{benefit}</span>
-                </li>
-              ))}
-            </ul>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export const SkillsCv = () => (
-  <SkillsWrapper skillsKey="skillsCv">
-    <GridItem className="col-span-10 sm:col-span-4">
-      <h2 className="text-4xl font-bold sm:text-6xl sm:font-normal">
-        Beruflicher Werdegang
-      </h2>
-    </GridItem>
-    <GridItem className="col-span-10 sm:col-span-6">
-      <p className="mb-8 text-xl sm:text-2xl">
-        Meine Karriere ist geprägt durch eine Mischung aus Freelance- und
-        Angestelltenverhältnissen, was mir eine breite Perspektive und tiefes
-        Verständnis für verschiedene Arbeitsumgebungen ermöglicht hat.
-      </p>
-    </GridItem>
-
-    <GridItem className="col-span-10">
-      <div className="relative space-y-8 border-l-2 border-gray-300 pl-8">
-        <CareerPhase
-          title="Erste Freelance-Phase"
-          period="2008 - 2011"
-          type="freelance"
-          description="Beginn meiner Karriere als selbstständiger Entwickler mit Fokus auf Web-Entwicklung und Software-Lösungen."
-          technologies={[
-            { name: 'PHP', icon: '/logos/php.svg', level: 'expert' },
-            { name: 'MySQL', icon: '/logos/mysql.svg', level: 'advanced' },
-            {
-              name: 'JavaScript',
-              icon: '/logos/javascript.svg',
-              level: 'advanced',
-            },
-          ]}
-          achievements={[
-            {
-              title: 'Entwicklung von E-Commerce Lösungen',
-              description:
-                'Implementierung mehrerer erfolgreicher Online-Shops mit individuellen Anforderungen',
-              impact:
-                'Steigerung der Online-Umsätze um durchschnittlich 40% bei den Kunden',
-            },
-          ]}
-          customerBenefits={[
-            'Erfahrung in der Entwicklung von skalierbaren Webanwendungen',
-            'Tiefes Verständnis für E-Commerce Anforderungen und Best Practices',
-            'Fähigkeit, komplexe technische Lösungen verständlich zu kommunizieren',
-          ]}
-        />
-        <CareerPhase
-          title="Angestelltenverhältnisse"
-          period="2011 - 2025"
-          type="employed"
-          description="14 Jahre Erfahrung in verschiedenen Unternehmen, wo ich tiefe Einblicke in Unternehmensstrukturen und Teamarbeit gewonnen habe."
-          technologies={[
-            { name: 'React', icon: '/logos/react.svg', level: 'expert' },
-            {
-              name: 'TypeScript',
-              icon: '/logos/typescript.svg',
-              level: 'expert',
-            },
-            { name: 'Node.js', icon: '/logos/nodejs.svg', level: 'advanced' },
-          ]}
-          achievements={[
-            {
-              title: 'Leitung von Entwicklerteams',
-              description:
-                'Verantwortlich für die technische Leitung und Mentoring von Entwicklerteams',
-              impact:
-                'Erfolgreiche Durchführung von 20+ Großprojekten mit internationalen Teams',
-            },
-          ]}
-          customerBenefits={[
-            'Umfassende Erfahrung in der Zusammenarbeit mit großen Unternehmen',
-            'Bewährte Methoden für effizientes Projektmanagement und Teamführung',
-            'Expertise in der Integration von modernen Technologien in bestehende Systeme',
-          ]}
-        />
-        <CareerPhase
-          title="Aktuelle Freelance-Phase"
-          period="2025 - heute"
-          type="freelance"
-          description="Zurück zur Selbstständigkeit mit dem Ziel, Unternehmen durch moderne Technologien und agile Methoden zu unterstützen."
-          technologies={[
-            { name: 'Next.js', icon: '/logos/nextjs.svg', level: 'expert' },
-            { name: 'AWS', icon: '/logos/aws.svg', level: 'advanced' },
-            { name: 'Docker', icon: '/logos/docker.svg', level: 'advanced' },
-          ]}
-          achievements={[
-            {
-              title: 'Moderne Web-Architekturen',
-              description:
-                'Entwicklung und Implementierung von skalierbaren Web-Anwendungen mit Fokus auf Performance und Benutzerfreundlichkeit',
-              impact:
-                'Reduzierung der Ladezeiten um 60% und Verbesserung der Conversion-Rate um 25%',
-            },
-          ]}
-          customerBenefits={[
-            'Kombination aus jahrelanger Unternehmenserfahrung und modernster Technologie-Expertise',
-            'Agile Entwicklung mit Fokus auf schnelle Time-to-Market',
-            'Individuelle Beratung und maßgeschneiderte Lösungen für Ihr Unternehmen',
-          ]}
-        />
-      </div>
-    </GridItem>
-
-    <GridItem className="col-span-10 mt-12">
-      <div className="flex flex-col justify-center gap-8 sm:flex-row">
-        <a
-          href="/docs/acun_guersoy_cv_2025.pdf"
-          target="_blank"
-          className="group flex items-center gap-4 rounded-lg border-2 border-gray-200 p-4 transition-all hover:border-blue-500"
-        >
-          <DocumentIcon className="size-8 text-blue-500 transition-transform group-hover:scale-110" />
-          <div>
-            <p className="font-bold group-hover:text-blue-500">
-              Lebenslauf als PDF
-            </p>
-            <p className="text-sm text-gray-600">
-              Detaillierte Version herunterladen
-            </p>
-          </div>
-        </a>
-        <a
-          href="https://www.linkedin.com/in/acun-g%C3%BCrsoy-83b8ab139/"
-          target="_blank"
-          className="group flex items-center gap-4 rounded-lg border-2 border-gray-200 p-4 transition-all hover:border-blue-500"
-        >
-          <div className="relative size-8">
-            <Image
-              fill
-              src="/logos/linkedin.svg"
-              alt="LinkedIn Logo"
-              className="object-contain"
-            />
-          </div>
-          <div>
-            <p className="font-bold group-hover:text-blue-500">
-              LinkedIn Profil
-            </p>
-            <p className="text-sm text-gray-600">Mein berufliches Netzwerk</p>
-          </div>
-        </a>
-      </div>
-    </GridItem>
-  </SkillsWrapper>
-)
