--- app/sections/Skills/index.ts
@@ -1,1 +1,0 @@
-export * from './Skills'
