+++ app/sections/index.ts
@@ -1,4 +1,3 @@
 export * from './Contact'
 export * from './Hero'
 export * from './Projects'
-export * from './Skills'
